Move Gazebo plugin code out of examples
