--- examples/gazebo_uav/gazebo_panoramic_camera.yaml
@@ -1,8 +1,0 @@
-%YAML:1.0
----
-unwrapper:
-    centre: [0.5, 0.5]
-    inner: 0
-    outer: 0.5
-    offsetDegrees: 0
-    flip: 0
