+++ include/robots/gazebo/uav_plugin.h
