+++ resources/panoramic_camera_parameters/gazebo_panoramic_camera.yaml
