+++ src/robots/gazebo/CMakeLists.txt
@@ -1,5 +1,6 @@
 cmake_minimum_required(VERSION 3.1)
 include(../../../cmake/bob_robotics.cmake)
 BoB_module(SOURCES tank.cc uav.cc
+           GAZEBO_PLUGINS tank_plugin.cc uav_plugin.cc
            BOB_MODULES robots
            EXTERNAL_LIBS gazebo)
