+++ src/robots/gazebo/tank_plugin.cc
