+++ src/robots/gazebo/uav_plugin.cc
@@ -20,7 +20,7 @@
 typedef SSIZE_T ssize_t;
 #endif
 
-#include "gazebo_quadcopter_plugin.h"
+#include "robots/gazebo/uav_plugin.h"
 
 // Gazebo includes
 #include <gazebo/common/Assert.hh>
